Move VS extensibility classes into namespace matching folder structure
