+++ src/NuGet.Clients/NuGet.VisualStudio.Implementation/Extensibility/VsNuGetFramework.cs
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace NuGet.VisualStudio
+namespace NuGet.VisualStudio.Implementation.Extensibility
 {
     public class VsNuGetFramework : IVsNuGetFramework
     {
