+++ src/NuGet.Clients/NuGet.VisualStudio.Implementation/Extensibility/VsPackageInstaller.cs
@@ -28,7 +28,7 @@ using NuGet.VisualStudio.Implementation.Resources;
 using NuGet.VisualStudio.Telemetry;
 using Task = System.Threading.Tasks.Task;
 
-namespace NuGet.VisualStudio
+namespace NuGet.VisualStudio.Implementation.Extensibility
 {
     [Export(typeof(IVsPackageInstaller))]
     [Export(typeof(IVsPackageInstaller2))]
@@ -379,7 +379,7 @@ namespace NuGet.VisualStudio
                         nameof(source));
                 }
 
-                var newSource = new Configuration.PackageSource(source);
+                var newSource = new PackageSource(source);
 
                 repo = _sourceRepositoryProvider.CreateRepository(newSource);
             }
@@ -426,7 +426,7 @@ namespace NuGet.VisualStudio
                 // Check if default package format is set to `PackageReference` and project has no
                 // package installed yet then upgrade it to `PackageReference` based project.
                 if (preferPackageReference &&
-                   (nuGetProject is MSBuildNuGetProject) &&
+                   nuGetProject is MSBuildNuGetProject &&
                    !(await nuGetProject.GetInstalledPackagesAsync(token)).Any() &&
                    await NuGetProjectUpgradeUtility.IsNuGetProjectUpgradeableAsync(nuGetProject, project, needsAPackagesConfig: false))
                 {
