+++ src/NuGet.Clients/NuGet.VisualStudio.Implementation/Extensibility/VsPackageMetadata.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using NuGet.Packaging.Core;
 
-namespace NuGet.VisualStudio
+namespace NuGet.VisualStudio.Implementation.Extensibility
 {
     internal class VsPackageMetadata : IVsPackageMetadata
     {
