+++ src/NuGet.Clients/NuGet.VisualStudio.Implementation/Extensibility/VsPackageMetadataComparer.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace NuGet.VisualStudio
+namespace NuGet.VisualStudio.Implementation.Extensibility
 {
     internal class VsPackageMetadataComparer : IEqualityComparer<IVsPackageMetadata>
     {
