+++ src/NuGet.Clients/NuGet.VisualStudio.Implementation/Extensibility/VsPackageProjectMetadata.cs
@@ -1,10 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
-using NuGet.ProjectManagement;
-
-namespace NuGet.VisualStudio
+namespace NuGet.VisualStudio.Implementation.Extensibility
 {
     internal class VsPackageProjectMetadata : IVsPackageProjectMetadata
     {
